move export tests
