+++ src/lib/import_export/file_type.hpp
@@ -4,7 +4,7 @@
 
 namespace opossum {
 
-enum class FileType : uint8_t { Csv, Tbl, Binary, Auto };
+enum class FileType { Csv, Tbl, Binary, Auto };
 
 FileType import_type_to_file_type(hsql::ImportType import_type);
